DataAccessObject erweitert, GoalDAO implementiert

+++ src/at/ac/tuwien/hci/ghost/data/entities/Entity.java
@@ -6,5 +6,5 @@ public abstract class Entity implements Serializable {
 	/** the serial version UID */
 	private static final long serialVersionUID = -5700235356938057111L;
 
-	public abstract int getID();
+	public abstract long getID();
 }
